channel lock

+++ src/Events/messageEvents/messageCreate/messageCreate.ts
@@ -39,6 +39,7 @@ const willis = (msg: Discord.Message) => {
 
 const eventDel = (msg: Discord.Message) => {
  if (msg.author?.id !== '1105195321732116530') return;
+ if (msg.channelId !== '298954459172700181') return;
 
  setTimeout(() => {
   if (msg) msg.delete();
